requirements.txt
Flask
requests
python-dotenv
pandas
plotly==5.14.1